backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
networkx
lancedb
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu
llama-cpp-python
pymupdf
sqlalchemy
aiosqlite
alembic
python-frontmatter
markdown-it-py
mdit-py-plugins
watchdog
huggingface-hub
pyarrow
pytest
pytest-asyncio
httpx
